Use single fixture file, cleanup from master
We've been having lots of trouble getting the fixture data correct. A lot of that is my fault by being too restrictive on the models we export. We'll just dump the whole DB and that'll make sure we don't have data integrity errors.

I also changed around the migrations because master had a new migration