Embed styleguide in edit page
I widened the constants that were previously being used so the tabs could fit and more content could fit on the screen. Unfortunately this blocks some of the controls for editing so we might have to play with how this fits together. This seems like an OK start.

Here's how things currently look:
![styleguide](https://user-images.githubusercontent.com/272675/42719343-ab3c8f20-86d9-11e8-8523-226c9c959a13.gif)